LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/delete-without-unwrapping-first-line-of-child-block_method=BackspaceKey&lineBreak=preformat-expected.txt


def
ghi

FAIL Backspace at abc\\n[\\n<div id="child">]def<br>ghi</div> assert_in_array: Backspace at abc\\n[\\n<div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> value "abc\n\n<div id=\"child\">def<br>ghi</div>" not in array ["abc\n<div id=\"child\">def<br>ghi</div>", "abc<div id=\"child\">def<br>ghi</div>"]
FAIL Typing "X" at abc\n[\n<div id="child">]def<br>ghi</div> assert_in_array: Typing "X" at abc\n[\n<div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> value "abc\n\n<div id=\"child\">def<br>ghi</div>" not in array ["abc\n<div id=\"child\">Xdef<br>ghi</div>", "abc<div id=\"child\">Xdef<br>ghi</div>"]
FAIL Backspace at \\n[abc\\n<div id="child">]def<br>ghi</div> assert_equals: Backspace at \\n[abc\\n<div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "\n<div id=\"child\">def<br>ghi</div>" but got "\nabc\n<div id=\"child\">def<br>ghi</div>"
FAIL Typing "X" at \n[abc\n<div id="child">]def<br>ghi</div> assert_equals: Typing "X" at \n[abc\n<div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> expected "\n<div id=\"child\">Xdef<br>ghi</div>" but got "\nabc\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at \\n[\\n<div id="child">]def<br>ghi</div> assert_equals: Backspace at \\n[\\n<div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "\n<div id=\"child\">def<br>ghi</div>" but got "\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Typing "X" at \n[\n<div id="child">]def<br>ghi</div> assert_equals: Typing "X" at \n[\n<div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> expected "\n<div id=\"child\">Xdef<br>ghi</div>" but got "\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at [\\n\\n<div id="child">]def<br>ghi</div> assert_equals: Backspace at [\\n\\n<div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Typing "X" at [\n\n<div id="child">]def<br>ghi</div> assert_equals: Typing "X" at [\n\n<div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> expected "<div id=\"child\">Xdef<br>ghi</div>" but got "\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at [abc\\n\\n<div id="child">]def<br>ghi</div> assert_equals: Backspace at [abc\\n\\n<div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "abc\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Typing "X" at [abc\n\n<div id="child">]def<br>ghi</div> assert_equals: Typing "X" at [abc\n\n<div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> expected "<div id=\"child\">Xdef<br>ghi</div>" but got "abc\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at abc\\n[\\n<div id="child">d]ef<br>ghi</div> assert_in_array: Backspace at abc\\n[\\n<div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected text in the <div> value "abc\n\n<div id=\"child\">def<br>ghi</div>" not in array ["abc\n<div id=\"child\">ef<br>ghi</div>", "abc<div id=\"child\">ef<br>ghi</div>"]
FAIL Typing "X" at abc\n[\n<div id="child">d]ef<br>ghi</div> assert_in_array: Typing "X" at abc\n[\n<div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> value "abc\n\n<div id=\"child\">def<br>ghi</div>" not in array ["abc\n<div id=\"child\">Xef<br>ghi</div>", "abc<div id=\"child\">Xef<br>ghi</div>"]
FAIL Backspace at \\n[abc\\n<div id="child">d]ef<br>ghi</div> assert_equals: Backspace at \\n[abc\\n<div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and the selected content in the <div> expected "\n<div id=\"child\">ef<br>ghi</div>" but got "\nabc\n<div id=\"child\">def<br>ghi</div>"
FAIL Typing "X" at \n[abc\n<div id="child">d]ef<br>ghi</div> assert_equals: Typing "X" at \n[abc\n<div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> expected "\n<div id=\"child\">Xef<br>ghi</div>" but got "\nabc\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at \\n[\\n<div id="child">d]ef<br>ghi</div> assert_equals: Backspace at \\n[\\n<div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected content in the <div> expected "\n<div id=\"child\">ef<br>ghi</div>" but got "\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Typing "X" at \n[\n<div id="child">d]ef<br>ghi</div> assert_equals: Typing "X" at \n[\n<div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> expected "\n<div id=\"child\">Xef<br>ghi</div>" but got "\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at [\\n\\n<div id="child">d]ef<br>ghi</div> assert_equals: Backspace at [\\n\\n<div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected content in the <div> expected "<div id=\"child\">ef<br>ghi</div>" but got "\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Typing "X" at [\n\n<div id="child">d]ef<br>ghi</div> assert_equals: Typing "X" at [\n\n<div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> expected "<div id=\"child\">Xef<br>ghi</div>" but got "\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at [abc\\n\\n<div id="child">d]ef<br>ghi</div> assert_equals: Backspace at [abc\\n\\n<div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected content in the <div> expected "<div id=\"child\">ef<br>ghi</div>" but got "abc\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Typing "X" at [abc\n\n<div id="child">d]ef<br>ghi</div> assert_equals: Typing "X" at [abc\n\n<div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> expected "<div id=\"child\">Xef<br>ghi</div>" but got "abc\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at abc\\n\\n<div id="child">[]def<br>ghi</div> assert_in_array: Backspace at abc\\n\\n<div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> value "abc\n\n<div id=\"child\">def<br>ghi</div>" not in array ["abc\n<div id=\"child\">def<br>ghi</div>", "abc<div id=\"child\">def<br>ghi</div>"]
FAIL Backspace at \\n\\n<div id="child">[]def<br>ghi</div> assert_equals: Backspace at \\n\\n<div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "\n<div id=\"child\">def<br>ghi</div>" but got "\n\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at \\n<div id="child">[]def<br>ghi</div> assert_equals: Backspace at \\n<div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "\n<div id=\"child\">def<br>ghi</div>"
FAIL Backspace at <b>abc\\n\\n</b></b><div id="child">[]def<br>ghi</div> assert_in_array: Backspace at <b>abc\\n\\n</b></b><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> (<b> should stay) value "<b>abc\n\n</b><div id=\"child\">def<br>ghi</div>" not in array ["<b>abc\n</b><div id=\"child\">def<br>ghi</div>", "<b>abc</b><div id=\"child\">def<br>ghi</div>"]
FAIL Backspace at <b>\\n</b><div id="child">[]def<br>ghi</div> assert_equals: Backspace at <b>\\n</b><div id="child">[]def<br>ghi</div> should delete only the preceding empty line (including the <b>) of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "<b>\n</b><div id=\"child\">def<br>ghi</div>"
